ml-server/requirements.txt
fastapi
uvicorn[standard]
transformers
scikit-learn
biopython
numpy
sentencepiece
protobuf
einops
huggingface_hub